@@ -7,13 +7,14 @@
     "@blockly/field-date": "^8.0.0",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
+    "@types/react-datepicker": "^4.19.4",
     "blockly": "^10.0.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
     "lucide-react": "^0.297.0",
-    "react": "^17.0.1",
+    "react": "^18.2.0",
     "react-datepicker": "^3.3.0",
-    "react-dom": "^17.0.1",
+    "react-dom": "^18.2.0",
     "react-syntax-highlighter": "^15.5.0",
     "tailwind-merge": "^2.1.0",
     "tailwindcss-animate": "^1.0.7"
@@ -49,8 +50,9 @@
   },
   "devDependencies": {
     "@types/node": "^20.8.6",
-    "@types/react": "^18.2.28",
-    "@types/react-dom": "^18.2.13",
+    "@types/react": "^18.2.43",
+    "@types/react-dom": "^18.2.17",
+    "@types/react-syntax-highlighter": "^15.5.11",
     "@vitejs/plugin-react": "^4.2.1",
     "autoprefixer": "^10.4.16",
     "postcss": "^8.4.31",
