@@ -26,31 +26,31 @@ import BlocklyComponent from "./BlocklyComponent";
 
 export default BlocklyComponent;
 
-const Block = (p) => {
+const Block = (p: any) => {
   const { children, ...props } = p;
   props.is = "blockly";
   return React.createElement("block", props, children);
 };
 
-const Category = (p) => {
+const Category = (p: any) => {
   const { children, ...props } = p;
   props.is = "blockly";
   return React.createElement("category", props, children);
 };
 
-const Value = (p) => {
+const Value = (p: any) => {
   const { children, ...props } = p;
   props.is = "blockly";
   return React.createElement("value", props, children);
 };
 
-const Field = (p) => {
+const Field = (p: any) => {
   const { children, ...props } = p;
   props.is = "blockly";
   return React.createElement("field", props, children);
 };
 
-const Shadow = (p) => {
+const Shadow = (p: any) => {
   const { children, ...props } = p;
   props.is = "blockly";
   return React.createElement("shadow", props, children);
