@@ -26,10 +26,10 @@
 
 import { javascriptGenerator } from "blockly/javascript";
 
-javascriptGenerator.forBlock["test_react_field"] = function (block) {
+javascriptGenerator.forBlock["test_react_field"] = function () {
   return "console.log('custom block');\n";
 };
 
-javascriptGenerator.forBlock["test_react_date_field"] = function (block) {
+javascriptGenerator.forBlock["test_react_date_field"] = function (block: any) {
   return "console.log(" + block.getField("DATE").getText() + ");\n";
 };
